Устранил ошибку разворота попапа, часть 2
