--- src/view/popup-comments.js
@@ -1,100 +1,0 @@
-import dayjs from 'dayjs';
-import relativeTime from 'dayjs/plugin/relativeTime.js';
-import {createElement} from '../utils.js';
-
-dayjs.extend(relativeTime);
-
-const createPopupCommentsTemplate = (length) => (
-  `<div class="film-details__bottom-container">
-    <section class="film-details__comments-wrap">
-      <h3 class="film-details__comments-title">
-        Comments
-        <span class="film-details__comments-count">${length}</span>
-      </h3>
-
-      <ul class="film-details__comments-list">
-
-      </ul>
-
-      <div class="film-details__new-comment">
-        <div class="film-details__add-emoji-label"></div>
-
-        <label class="film-details__comment-label">
-          <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-        </label>
-
-        <div class="film-details__emoji-list">
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-          <label class="film-details__emoji-label" for="emoji-smile">
-            <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-          </label>
-
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-          <label class="film-details__emoji-label" for="emoji-sleeping">
-            <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-          </label>
-
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-          <label class="film-details__emoji-label" for="emoji-puke">
-            <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-          </label>
-
-          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-          <label class="film-details__emoji-label" for="emoji-angry">
-            <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-          </label>
-        </div>
-      </div>
-    </section>
-  </div>`
-);
-
-export const generateCommentsList = (comments = {}) => {
-  const commentsList = document.querySelector('.film-details__comments-list');
-  commentsList.innerHTML = '';
-  for (const commentary of comments) {
-    const {author, comment, date, emotion} = commentary;
-
-    const timePassed = dayjs(date).fromNow();
-
-    const elementList = `
-    <li class="film-details__comment">
-      <span class="film-details__comment-emoji">
-        <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-smile">
-      </span>
-      <div>
-        <p class="film-details__comment-text">${comment}</p>
-        <p class="film-details__comment-info">
-          <span class="film-details__comment-author">${author}</span>
-          <span class="film-details__comment-day">${timePassed}</span>
-          <button class="film-details__comment-delete">Delete</button>
-        </p>
-      </div>
-    </li>`;
-    commentsList.insertAdjacentHTML('beforeEnd', elementList);
-  }
-};
-
-export default class PopupComments {
-  constructor(length) {
-    this._length = length;
-    this._element = null;
-  }
-
-  getTemplate() {
-
-    return createPopupCommentsTemplate(this._length);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-}
