+++ src/view/popup.js
@@ -1,14 +1,21 @@
 import {isClassNamePopup, createElement} from '../utils.js';
+import dayjs from 'dayjs';
+import relativeTime from 'dayjs/plugin/relativeTime.js';
+
+dayjs.extend(relativeTime);
 
 const isGenre = (array) => array.length > 1 ? 'Genres' : 'Genre';
 
-const createPopupFilmTemplate = (film = {}) => {
-  const {filmInfo, userDetails} = film;
+const createPopupTemplate = (card) => {
+  const {film} = card;
+  const {filmInfo, userDetails, comments} = film;
   const {title, totalRating, releaseFilm, genre, poster, description, ageRating, alternativeTitle, writers, director, actors, runtime} = filmInfo;
   const {date, releaseCountry} = releaseFilm;
   const {watchList, alreadyWatched, favorite} = userDetails;
 
-  return `<div class="film-details__top-container">
+  return `<section class="film-details">
+    <form class="film-details__inner" action="" method="get">
+    <div class="film-details__top-container">
       <div class="film-details__close">
         <button class="film-details__close-btn" type="button">close</button>
       </div>
@@ -59,7 +66,7 @@ const createPopupFilmTemplate = (film = {}) => {
             <tr class="film-details__row">
               <td class="film-details__term">${isGenre(genre)}</td>
               <td class="film-details__cell film-details__cell-genre">
-
+                    ${genre.join(', ')}
               </td>
             </tr>
           </table>
@@ -73,35 +80,92 @@ const createPopupFilmTemplate = (film = {}) => {
         <button type="button" class="film-details__control-button film-details__control-button--watched ${isClassNamePopup(watchList)}" id="watched " name="watched">Already watched</button>
         <button type="button" class="film-details__control-button film-details__control-button--favorite ${isClassNamePopup(favorite)}" id="favorite" name="favorite">Add to favorites</button>
       </section>
-    </div>`;
+    </div>
+
+    <div class="film-details__bottom-container">
+    <section class="film-details__comments-wrap">
+      <h3 class="film-details__comments-title">
+        Comments
+        <span class="film-details__comments-count">${comments.length}</span>
+      </h3>
+
+      <ul class="film-details__comments-list">
+
+      </ul>
+
+      <div class="film-details__new-comment">
+        <div class="film-details__add-emoji-label"></div>
+
+        <label class="film-details__comment-label">
+          <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
+        </label>
+
+        <div class="film-details__emoji-list">
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
+          <label class="film-details__emoji-label" for="emoji-smile">
+            <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
+          <label class="film-details__emoji-label" for="emoji-sleeping">
+            <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
+          <label class="film-details__emoji-label" for="emoji-puke">
+            <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
+          </label>
+
+          <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
+          <label class="film-details__emoji-label" for="emoji-angry">
+            <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
+          </label>
+        </div>
+      </div>
+    </section>
+    </div>
+    </form>
+  </section>`;
 };
 
-export const generateGenreList = (array = {}) => {
-  const genreList = document.querySelector('.film-details__cell-genre');
-  genreList.innerHTML = '';
-  array.filmInfo.genre.forEach((item) => {
-    const elementList = document.createElement('span');
-    elementList.classList.add('film-details__genre');
-    elementList.textContent = item;
-    genreList.appendChild(elementList);
-  });
+export const generateCommentsList = (comments = {}) => {
+  const commentsList = document.querySelector('.film-details__comments-list');
+  for (const commentary of comments) {
+    const {author, comment, date, emotion} = commentary;
+
+    const timePassed = dayjs(date).fromNow();
+
+    const elementList = `
+    <li class="film-details__comment">
+      <span class="film-details__comment-emoji">
+        <img src="./images/emoji/${emotion}.png" width="55" height="55" alt="emoji-smile">
+      </span>
+      <div>
+        <p class="film-details__comment-text">${comment}</p>
+        <p class="film-details__comment-info">
+          <span class="film-details__comment-author">${author}</span>
+          <span class="film-details__comment-day">${timePassed}</span>
+          <button class="film-details__comment-delete">Delete</button>
+        </p>
+      </div>
+    </li>`;
+    commentsList.insertAdjacentHTML('beforeEnd', elementList);
+  }
 };
 
-export default class PopupFilm {
+export default class Popup {
   constructor(film) {
     this._film = film;
     this._element = null;
   }
 
   getTemplate() {
-
-    return createPopupFilmTemplate(this._film);
+    return createPopupTemplate(this._film);
   }
 
   getElement() {
     if (!this._element) {
       this._element = createElement(this.getTemplate());
-
     }
 
     return this._element;
